+++ NEWS.md
@@ -1,6 +1,6 @@
 # explore 1.3.2
 
-* change default color from grey to #C1D5E6 (grey/blue)
+* change default color from grey to "#ADD8E6" (grey/blue)
 
 # explore 1.3.1